AI/requirements.txt
pypdf==5.0.1
python-docx==1.1.2
spacy==3.7.5
sentence-transformers==3.0.1
scikit-learn==1.5.1
scipy==1.13.1
django==5.0.7
djangorestframework==3.15.2
celery==5.4.0
redis==5.0.7
python-dotenv==1.0.1
pandas==2.2.2